+++ NEWS.md
@@ -3,6 +3,7 @@
   - `Implies` for logical implies
   - `is_sorted` and `isntSorted` for assertions about sorted atomic vectors
   - `fmatchp`, `finp` experimental parallel hashing functions
+  - `minmax` multithreaded function of `c(min(x), max(x))`
 
 ## hutilscpp 0.5.2
 * Attempt fix of UBSAN error in which_first